refactor: rename trainer modules and update imports for consistency
