+++ tests/trainers/test_hf_trainer.py
@@ -25,13 +25,13 @@ from omegaconf import DictConfig
 
 
 @pytest.fixture
-def hf_bert_trainer():
+def hf_bert():
     from supertrainer.trainers.mllm_trainer import BERTTrainer
 
     return BERTTrainer(DictConfig({"classes": ["a", "b", "c"]}))
 
 
-def test_config(hf_bert_trainer):
-    assert hf_bert_trainer.config.num_classes == 3
-    assert hf_bert_trainer.config.class2id == {"a": 0, "b": 1, "c": 2}
-    assert hf_bert_trainer.config.id2class == {0: "a", 1: "b", 2: "c"}
+def test_config(hf_bert):
+    assert hf_bert.config.num_classes == 3
+    assert hf_bert.config.class2id == {"a": 0, "b": 1, "c": 2}
+    assert hf_bert.config.id2class == {0: "a", 1: "b", 2: "c"}
